Call consensus restructuring (#11)
* Consensus read calling restructured

* Minor change

* Revised structure

* cargo-fmt
